+++ platform/src/sdk/models/operations/updateconnectionauthorization.ts
@@ -37,6 +37,12 @@ export class UpdateConnectionAuthorizationResponse extends SpeakeasyBase {
     contentType: string;
 
     /**
+     * Your API request was not properly authorized.
+     */
+    @SpeakeasyMetadata()
+    errorMessage?: shared.ErrorMessage;
+
+    /**
      * HTTP response status code for this operation
      */
     @SpeakeasyMetadata()